hashi::mpc

Struct MpcManager

pub struct MpcManager {

    pub party_id: PartyId,
    pub address: Address,
    pub mpc_config: MpcConfig,
    pub session_id: SessionId,
    pub encryption_key: PrivateKey<EncryptionGroupElement>,
    pub signing_key: Bls12381PrivateKey,
    pub committee: Committee,
    pub previous_committee: Option<Committee>,
    pub previous_nodes: Option<Nodes<EncryptionGroupElement>>,
    pub previous_threshold: Option<u16>,
    pub source_epoch: u64,
    pub batch_size_per_weight: u16,
    pub dealer_outputs: HashMap<DealerOutputsKey, PartialOutput>,
    pub dkg_messages: HashMap<Address, Message>,
    pub rotation_messages: HashMap<Address, RotationMessages>,
    pub nonce_messages: HashMap<Address, NonceMessage>,
    pub message_responses: HashMap<Address, SendMessagesResponse>,
    pub complaints_to_process: HashMap<ComplaintsToProcessKey, Complaint>,
    pub complaint_responses: HashMap<(Address, ProtocolTypeIndicator), ComplaintResponses>,
    pub public_messages_store: Box<dyn PublicMessagesStore>,
    pub dealer_nonce_outputs: BTreeMap<Address, ReceiverOutput>,
    /* private fields */

}

Fields
party_id: PartyIdaddress: Addressmpc_config: MpcConfigsession_id: SessionIdencryption_key: PrivateKey<EncryptionGroupElement>signing_key: Bls12381PrivateKeycommittee: Committeeprevious_committee: Option<Committee>previous_nodes: Option<Nodes<EncryptionGroupElement>>previous_threshold: Option<u16>source_epoch: u64

The epoch from which to read previous messages during reconstruction.

batch_size_per_weight: u16dealer_outputs: HashMap<DealerOutputsKey, PartialOutput>dkg_messages: HashMap<Address, Message>rotation_messages: HashMap<Address, RotationMessages>nonce_messages: HashMap<Address, NonceMessage>message_responses: HashMap<Address, SendMessagesResponse>complaints_to_process: HashMap<ComplaintsToProcessKey, Complaint>complaint_responses: HashMap<(Address, ProtocolTypeIndicator), ComplaintResponses>public_messages_store: Box<dyn PublicMessagesStore>dealer_nonce_outputs: BTreeMap<Address, ReceiverOutput>

Must be BTreeMap so that all nodes iterate outputs in the same deterministic order when constructing Presignatures.

Implementations

impl MpcManager

pub fn new( address: Address, committee_set: &CommitteeSet, session_id: SessionId, encryption_key: PrivateKey<EncryptionGroupElement>, signing_key: Bls12381PrivateKey, public_message_store: Box<dyn PublicMessagesStore>, allowed_delta: u16, chain_id: &str, weight_divisor: Option<u16>, batch_size_per_weight: u16, test_corrupt_shares_for: Option<Address>, ) -> MpcResult<Self>

pub fn handle_send_messages_request( &mut self, sender: Address, request: &SendMessagesRequest, ) -> MpcResult<SendMessagesResponse>

pub fn handle_retrieve_messages_request( &self, request: &RetrieveMessagesRequest, ) -> MpcResult<RetrieveMessagesResponse>

pub fn handle_complain_request( &mut self, request: &ComplainRequest, ) -> MpcResult<ComplaintResponses>

pub fn handle_get_public_mpc_output_request( &self, request: &GetPublicMpcOutputRequest, ) -> MpcResult<GetPublicMpcOutputResponse>

pub async fn run_dkg( mpc_manager: &Arc<RwLock<Self>>, p2p_channel: &impl P2PChannel, tob_channel: &mut impl OrderedBroadcastChannel<CertificateV1>, ) -> MpcResult<MpcOutput>

pub async fn run_key_rotation( mpc_manager: &Arc<RwLock<Self>>, previous_certificates: &[CertificateV1], p2p_channel: &impl P2PChannel, ordered_broadcast_channel: &mut impl OrderedBroadcastChannel<CertificateV1>, ) -> MpcResult<MpcOutput>

pub async fn run_nonce_generation( mpc_manager: &Arc<RwLock<Self>>, batch_index: u32, p2p_channel: &impl P2PChannel, tob_channel: &mut impl OrderedBroadcastChannel<CertificateV1>, ) -> MpcResult<Vec<ReceiverOutput>>

pub fn reconstruct_presignatures( &self, batch_index: u32, certs: &[(Address, DealerSubmissionV1)], ) -> MpcResult<NonceReconstructionOutcome>

pub async fn fetch_public_mpc_output_from_quorum( mpc_manager: &Arc<RwLock<Self>>, p2p_channel: &impl P2PChannel, previous_committee_threshold: u64, ) -> MpcResult<PublicMpcOutput>