hashi::onchain

Module types

Usable definitions of the onchain state of hashi

Structs

Coin

CommitteeSet

Config

DepositConfirmationMessage

Message signed by the committee to confirm a deposit. Mirrors Move deposit::DepositConfirmationMessage.

DepositRequest

Rust version of the Move hashi::deposit_queue::DepositRequest type.

DepositRequestQueue

Hashi

MemberInfo

MetadataCap

OutputUtxo

Rust version of the Move hashi::withdrawal_queue::OutputUtxo type.

Proposal

A proposal stored in the proposals bag

Proposals

Proposals bag - stores governance proposals by ID

Treasury

TreasuryCap

UpgradeCap

Rust version of the Move sui::package::UpgradeCap type.

Utxo

UtxoId

txid:vout

UtxoPool

UtxoRecord

Rust version of the Move hashi::utxo_pool::UtxoRecord type.

WithdrawalRequest

Rust version of the Move hashi::withdrawal_queue::WithdrawalRequest type.

WithdrawalRequestQueue

WithdrawalTransaction

Rust version of the Move hashi::withdrawal_queue::WithdrawalTransaction type.

Enums

ConfigValue

Rust version of the Move hashi::config_value::Value type.

ProposalType

The type of proposal data stored in a Proposal<T>

WithdrawalStatus

Rust version of the Move hashi::withdrawal_queue::WithdrawalStatus enum.

Constants

DEFAULT_MPC_MAX_FAULTY_IN_BASIS_POINTS

Default MPC max faulty parties in basis points. Mirrors DEFAULT_MAX_FAULTY_IN_BASIS_POINTS in mpc_config.move.

DEFAULT_MPC_THRESHOLD_IN_BASIS_POINTS

Default MPC threshold in basis points. Mirrors DEFAULT_THRESHOLD_IN_BASIS_POINTS in mpc_config.move.

DEFAULT_MPC_WEIGHT_REDUCTION_ALLOWED_DELTA

Default allowed delta for weight reduction. Mirrors DEFAULT_WEIGHT_REDUCTION_ALLOWED_DELTA in mpc_config.move.